Product page: Generate product card
created displayProduct.js
