+++ product.html
@@ -45,7 +45,7 @@
         </div>
       </div>
     </article>
-    <div class="product-card">
+    <div class="product-card" id="product-card">
       <div class="product-container">
         <div class="image">
           <img src="./img/image 1.jpg" alt="Product image">
